+++ src/Menu.tsx
@@ -1,5 +1,5 @@
 //a component is a function that returns a jsx/tsx template, and that function
-//is exported at the bottom of the
+//is exported at the bottom of the function
 
 function Menu() {
     return (
